Remove db/util.go and add txMeta methods to infrastructure
