--- cmd/soroban-rpc/internal/db/util.go
@@ -1,107 +1,0 @@
-package db
-
-import (
-	"github.com/stellar/go/network"
-	"github.com/stellar/go/xdr"
-)
-
-// Methods for creating test txMeta
-
-const (
-	txMetaV                 = 3
-	ledgerCloseTimeConstant = 100
-	feeCharged              = 100
-)
-
-func CreateTxMeta(acctSeq uint32, successful bool) xdr.LedgerCloseMeta {
-	envelope := txEnvelope(acctSeq)
-	txProcessing := []xdr.TransactionResultMeta{
-		{
-			TxApplyProcessing: xdr.TransactionMeta{
-				V:          txMetaV,
-				Operations: &[]xdr.OperationMeta{},
-				V3:         &xdr.TransactionMetaV3{},
-			},
-			Result: xdr.TransactionResultPair{
-				TransactionHash: txHash(acctSeq),
-				Result:          transactionResult(successful),
-			},
-		},
-	}
-	components := []xdr.TxSetComponent{
-		{
-			Type: xdr.TxSetComponentTypeTxsetCompTxsMaybeDiscountedFee,
-			TxsMaybeDiscountedFee: &xdr.TxSetComponentTxsMaybeDiscountedFee{
-				BaseFee: nil,
-				Txs:     []xdr.TransactionEnvelope{envelope},
-			},
-		},
-	}
-
-	return xdr.LedgerCloseMeta{
-		V: 1,
-		V1: &xdr.LedgerCloseMetaV1{
-			LedgerHeader: xdr.LedgerHeaderHistoryEntry{
-				Header: xdr.LedgerHeader{
-					ScpValue: xdr.StellarValue{
-						CloseTime: xdr.TimePoint(LedgerCloseTime(acctSeq)),
-					},
-					LedgerSeq: xdr.Uint32(acctSeq),
-				},
-			},
-			TxProcessing: txProcessing,
-			TxSet: xdr.GeneralizedTransactionSet{
-				V: 1,
-				V1TxSet: &xdr.TransactionSetV1{
-					PreviousLedgerHash: xdr.Hash{1},
-					Phases: []xdr.TransactionPhase{{
-						V:            0,
-						V0Components: &components,
-					}},
-				},
-			},
-		},
-	}
-}
-
-func txHash(acctSeq uint32) xdr.Hash {
-	envelope := txEnvelope(acctSeq)
-	hash, err := network.HashTransactionInEnvelope(envelope, network.FutureNetworkPassphrase)
-	if err != nil {
-		panic(err)
-	}
-	return hash
-}
-
-func txEnvelope(acctSeq uint32) xdr.TransactionEnvelope {
-	envelope, err := xdr.NewTransactionEnvelope(xdr.EnvelopeTypeEnvelopeTypeTx, xdr.TransactionV1Envelope{
-		Tx: xdr.Transaction{
-			Fee:           1,
-			SeqNum:        xdr.SequenceNumber(acctSeq),
-			SourceAccount: xdr.MustMuxedAddress("MA7QYNF7SOWQ3GLR2BGMZEHXAVIRZA4KVWLTJJFC7MGXUA74P7UJVAAAAAAAAAAAAAJLK"),
-		},
-	})
-	if err != nil {
-		panic(err)
-	}
-	return envelope
-}
-
-func transactionResult(successful bool) xdr.TransactionResult {
-	code := xdr.TransactionResultCodeTxBadSeq
-	if successful {
-		code = xdr.TransactionResultCodeTxSuccess
-	}
-	opResults := []xdr.OperationResult{}
-	return xdr.TransactionResult{
-		FeeCharged: feeCharged,
-		Result: xdr.TransactionResultResult{
-			Code:    code,
-			Results: &opResults,
-		},
-	}
-}
-
-func LedgerCloseTime(ledgerSequence uint32) int64 {
-	return int64(ledgerSequence)*25 + ledgerCloseTimeConstant
-}
